pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "llama-index>=0.12.33",
    "llama-index-embeddings-huggingface>=0.5.3",
    "llama-index-llms-deepseek>=0.1.1",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
]
